+++ src/pages/mod.rs
@@ -1,0 +1,6 @@
+//! Based on: https://github.com/yewstack/yew/blob/yew-v0.19.3/examples/router/src/pages/mod.rs
+//!
+
+pub mod about;
+pub mod landing;
+pub mod page_not_found;
